requirements.txt
tensorflow-cpu
pandas
numpy
mlflow
